requirements.txt
anthropic==0.49.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
pgvector==0.3.6
neo4j==5.27.0
requests==2.32.3
pydantic==2.10.4
slowapi==0.1.9
python-dotenv==1.0.1
